Quick & Dirty way to find the critical path
If you want to use this method, you need a mxm matrix (at least on
paper) that corresponds to the grid of the maze (so 5x5 for round 1, 6x6
for round 2, 7x7 for round 3, i think)

the matrix is filled with numbers that index the points.  Each time the
robot enters a square, it sends the square's index using setNextStep(i).

The function will update the critical path depending on if we were in
that square just now or not.

When it's time to follow the critical path, use getNextStep to find out
which way to go each time.